moved protoParser to main directory
