+++ tsconfig.json
@@ -23,6 +23,6 @@
     /* Advanced Options */
     "forceConsistentCasingInFileNames": true /* Disallow inconsistently-cased references to the same file. */
   },
-  "include": ["./src/"],
+  "include": ["./src/", "protoParser.js"],
   "exclude": ["node_modules", "dist", "build", "coverage", "mocks"]
 }
